start layout editions graph
