+++ src/pages/main_dash.py
@@ -9,6 +9,6 @@ load_figure_template([config.dark_theme, config.light_theme])
 
 app = DashProxy(transforms=[LogTransform()],
                 external_stylesheets=[dbc.themes.CYBORG, dbc_css],
-                meta_tags=[{'name': 'viewport', 'content': 'width=device-width'}],
+                # meta_tags=[{'name': 'viewport', 'content': 'width=device-width'}],
                 # suppress_callback_exceptions=True
                 )
